Milestone 1: Prelinked kernel, multi-core, ACPI Platform Expert (#552)
* Fix return value in +[NSDrawer visibleEdgeWithPreferredEdge:parentWindow:drawerWindow:] (#541)
* Switch to dyld-750.6
* Fix libunwind dependency and set library versions to Monterey
* Kernel now links!
* Build pthread.kext
* Add a simple libCrashReporterClient to satisfy ABI
* Add enough of math.h to satisfy our needs
* Move Foundation headers for easier building
* Add cmake file for Foundation
* Start building kexts for PE
* Add Apple objc4 source
* Start on building libobjc, Foundation, CF for kext_tools
* Generate dyld platform version constants in dyld_priv.h
* Add more mach headers to SDK, C++ CRClient, revert BSD changes to libxpc
* Add stdatomic.h, libnv and libsbuf from FreeBSD
* Build GCD, objc4, libxpc
* Build libsystem_asl
* explicit bash instead of sh (#546)
* Correct some clean build issues
* Foundation now builds (not linked yet)
* Add sys headers to System.fmwk PrivateHeaders
* Import Apple CFNetwork 4.0 (2004)
* Nearly done porting Foundation to Apple objc + libSystem improvements
* Add mDNSResponder and re-org BSD folder
* Add DirectoryService from tag DirectoryService-655
* Add OpenDirectory from tag OpenDirectory-146
* Add configd from tag configd-1061.141.1
* Build host rpcgen to default toolchain
* Add rpcsvc. Generate headers to SDK.
* Refactor to post-10.12 sysdir API
* Build CoreFoundation (aside from sysdir)
* Add removefile to libsystem
* Build libcopyfile for libSystem
* Convert to standard atomic ops
* Build libsystem_info (some symbols undefined)
* Finally got the right symbol suffixing for libc!
* Implement stubs for libsystem_trace
* Implement xpc_dictionary_get_array and xpc_dictionary_get_dictionary
* Clean deprecated cruft
* Add missing sleep.c to FreeBSD.a
* Build libsystem_dnssd from mDNSResponder
* Start implementing CoreUtils.framework and internal SDK
* Moved out of date docs aside
* Build libsystem_dnssd and CoreUtils.framework skeleton
* Add lzvndecode tool (optional build)
* Update Platform and Boot args lists
* WIP on kext_tools build
* Add functions to host libmacho
* Log kernel build to console
* Fix a few properties
* Clean up old FreeBSD tree
* Add plktool. Revert HOST_BUILD changes to IOKit, kext_tools
* Remove swap.c - it is in libstuff too
* Start on makefile for kext project and IOKit userspace
* plktool linked with kxld!
* Finished first draft of plktool.
* plktool is working, sans compression
* Add the prelinkedkernel header to our output
* Refactor and fix prelinked kernel tool. (It works)

This commit splits my original plktool.c draft into multiple files,
formats code for readability, removes tons of cruft, adds missing
structures and functions, and fixes many bugs.

Most importantly, this commit provides a working Linux tool that
can create a prelinked kernel (kernelcache) that boots in my
qemu VM.

* Build private extension to System.kext
* Build IOACPIFamily kext
* Handle kext dependencies during link
* Fix writing of linked kexts to cache
* Add our PE kext infodict to kernelcache __BUILTIN segment

Our PE class needs to be initialized before IOKit matching starts. By
embedding the Info.plist into this special kernel segment, we get XNU
to add the IOKit personalities for the kext into the catalogue early.

* Include __BUILTIN segment for prelink
* Call C++ static constructors before _start
* Add special handling for our PE kext

Our PE kext needs to be initialized early to become the new
Platform Expert class. When the PE kext matches an IOKit
provider class, initialize the C++ runtime for the kext before
further matching.

* Merged with AppleI386GenericPlatform.
* Work in progress on PE kext
* Update kext build targets
* Bring up IOCPUs and APICs. Renamed files.

---------

+++ Frameworks/Foundation/NSURL/NSURLProtocol_http.m
@@ -347,7 +347,7 @@ NSLog(@"parse error %d %o",__LINE__,code);
 NSLog(@"zero chunk");
        }
        else {
-NSLog(@"chunk=%d",_chunkSize);
+NSLog(@"chunk=%d",(long)_chunkSize);
         _state=STATE_waitingForChunkCompletion;
         rangeAction=advanceLocationToNext;
        }
